[Chore] Mument Card Databinding Change
47
